Create Track Retweet Count
This project is to follow one specific user on Twitter. I connect to the tweepy api and then use the user_timeline command to fetch the most recent tweet on twitter. The id of the tweet is saved in a list and the retweet count is tracked based on the scheduled looped.